divide the logic to classes
